Refactor dependencies and test command in python-package.yml
In this commit, I refactored the dependencies installation and test command in the python-package.yml workflow file.

Instead of installing uvloop separately, I combined it with the other dependencies (pytest and flake8) in a single pip install command.

I also commented out the linting step using flake8, as it was not needed for now.

Finally, I replaced the "unittest" command with "pytest" for running the tests.